+++ R/jsBasicGadget.R
@@ -27,7 +27,7 @@ jsBasicGadget <- function(data) {
   change.vnlist = list(c(" ", "_"), c("=<", "_le_"), c("=>", "_ge_"), c("=", "_eq_"), c("\\(", "_open_"), c("\\)", "_close_"), c("%", "_percent_"), c("-", "_"), c("/", "_"),
                        c("\r\n", "_"), c(",", "_comma_"))
 
-  out <- data.table(data)
+  out <- data.table(data, check.names = T)
 
 
   ## Initial variable name